+++ src/components/Timer/DismissTimer.js
@@ -4,7 +4,7 @@ import Typography from '@material-ui/core/Typography';
 import { makeStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
 import {PRISMIC_GOOFY_BG, PRISMIC_COOL_BG, PRISMIC_BOO_BG} from '../../utils/constants';
-import {useRenderCounts, useCmsAsset} from '../../hooks/api-hooks';
+import {useRenderCounts, useCmsAsset, useWhatsapplogger} from '../../hooks/api-hooks';
 
 
 const useStyles = makeStyles(theme => ({
@@ -15,10 +15,12 @@ const useStyles = makeStyles(theme => ({
 
 const DismissTimer  = ({dismissMsg, setdismissMsg, anchorEl, setAnchorEl, timerdisplay, setTimerdisplay}) => {    
     useRenderCounts('DismissTimer.js');
+  const [log, setLogs] = useWhatsapplogger({});
+
     const [goofy, setGoofy] = useState('');
     const [cool, setCool] = useState('');
     const [boo, setBoo] = useState('');
-    const [emojis, setEmojis] = useState({})
+    // const [emojis, setEmojis] = useState({})
     const assets = useCmsAsset(PRISMIC_GOOFY_BG, PRISMIC_COOL_BG, PRISMIC_BOO_BG);
     useEffect(() => {
         if (assets.length) {
@@ -45,6 +47,11 @@ const DismissTimer  = ({dismissMsg, setdismissMsg, anchorEl, setAnchorEl, timerd
     const classes = useStyles();
     const onClose1 = (event) => {
         setAnchorEl(event.currentTarget);
+        // Whatsapp Logger
+        setLogs({
+            action: "Dismiss",
+            message: `just attempted for dismissing the timer`
+        });
         const timeleftEl = document.querySelector('.timerComp .MuiSnackbarContent-message div span:nth-child(1)').innerText.split(' ');
         const minsorsecs = timeleftEl[1];
         const minsorsecsVal = parseInt(timeleftEl[0]);
@@ -79,28 +86,38 @@ const DismissTimer  = ({dismissMsg, setdismissMsg, anchorEl, setAnchorEl, timerd
     const open = Boolean(anchorEl);
     const id = open ? 'simple-popover' : null;
     return (
-        <>
-            <Button color="secondary" size="small" onClick={onClose1} style={{marginTop:`${mtop}`}}>
-            {(dismissMsg[0] === 'Dismiss')?dismissMsg[0]:null}{dismissMsg[0] !== 'Dismiss' && <img src={dismissMsg[1]} width="60" height="60"/>}
-            </Button>
-            <Popover
-                id={id}
-                open={open}
-                anchorEl={anchorEl}
-                onClose={handleClose}
-                anchorOrigin={{
-                vertical: 'bottom',
-                horizontal: 'center',
-                }}
-                transformOrigin={{
-                vertical: 'top',
-                horizontal: 'center',
-                }}
-            >
-                <Typography className={classes.typography}>{dismissMsg[0]}</Typography>
-            </Popover>
-        </>
-    )
+      <>
+        <Button
+          color="secondary"
+          size="small"
+          onClick={onClose1}
+          style={{ marginTop: `${mtop}` }}
+        >
+          {dismissMsg[0] === "Dismiss" ? dismissMsg[0] : null}
+          {dismissMsg[0] !== "Dismiss" && (
+            <img src={dismissMsg[1]} width="60" height="60" />
+          )}
+        </Button>
+        <Popover
+          id={id}
+          open={open}
+          anchorEl={anchorEl}
+          onClose={handleClose}
+          anchorOrigin={{
+            vertical: "bottom",
+            horizontal: "center"
+          }}
+          transformOrigin={{
+            vertical: "top",
+            horizontal: "center"
+          }}
+        >
+          <Typography className={classes.typography}>
+            {dismissMsg[0]}
+          </Typography>
+        </Popover>
+      </>
+    );
 }
 
 export default DismissTimer;